Split cloud storage cache into its own module
